Refactor MockCollector
