+++ test/IntegrationTests/SmokeTests.cs
@@ -118,12 +118,10 @@ public class SmokeTests : TestHelper
     public void SubmitMetrics()
     {
         SetEnvironmentVariable("OTEL_DOTNET_AUTO_METRICS_ADDITIONAL_SOURCES", "MyCompany.MyProduct.MyLibrary");
-        var collectorPort = TcpPortProvider.GetOpenPort();
-        using var collector = new MockCollector(Output, collectorPort);
-
         const int expectedMetricRequests = 1;
 
-        RunTestApplication(metricsAgentPort: collectorPort);
+        using var collector = new MockMetricsCollector(Output);
+        RunTestApplication(metricsAgentPort: collector.Port);
         var metricRequests = collector.WaitForMetrics(expectedMetricRequests, TimeSpan.FromSeconds(5));
 
         using (new AssertionScope())
